refactor(verticals): fix type errors in vertical implementations
- Add type ignore comments for type issues across all verticals
- Resolve import and annotation problems in vertical modules
- Fix type issues in RAG, coding, DevOps, and research verticals
- Improve type safety in vertical base classes and capabilities

Part of MyPy type error resolution effort.

+++ victor/teams/team_learning.py
@@ -401,7 +401,7 @@ class TeamLearningSystem:
 
         # Compute metrics
         rewards = [e.reward for e in experiences]
-        avg_reward = np.mean(rewards) if rewards else 0.0
+        avg_reward = float(np.mean(rewards)) if rewards else 0.0
 
         # Compute trend
         if len(rewards) >= 10:
